fix: fail with a friendly error on unsupported Node.js versions
On Node.js versions older than the engines range, xst crashed deep in
the import chain of its dependencies (SyntaxError from execa) without
any hint at the cause.

The entry point is split: cli.js is now a thin launcher whose static
import graph contains only node: built-ins and semver (CJS, parses on
ancient Node.js). It checks process.versions.node against the engines
range read from package.json, prints a friendly error to stderr and
exits with code 1 on unsupported versions. Only then is the actual CLI
(app.js, the former cli.js content) loaded via dynamic import.

Closes #291
