+++ cli.js
@@ -1,86 +1,34 @@
 #!/usr/bin/env node
 
-import yargs from 'yargs'
-import chalk from 'chalk'
-import { commands } from './commands/index.js'
-import { handleError } from './utility/errors.js'
-import { hideBin } from 'yargs/helpers'
-import { readConnection } from './utility/connection.js'
-import { configure } from './utility/configure.js'
+// This launcher must parse and run on unsupported Node.js versions to be
+// able to fail with a friendly error message (see #291). Its static import
+// graph must contain no third-party code except 'semver' (CJS, runs on
+// ancient Node.js). The actual CLI is loaded via dynamic import below.
+import { readFileSync } from 'node:fs'
+import { isSupported } from './utility/node-version.js'
 
-// colored terminal output
-const bb = chalk.blackBright
-const ma = chalk.magenta
-const rb = chalk.redBright
-const yl = chalk.yellow
+const { engines } = JSON.parse(
+  readFileSync(new URL('./package.json', import.meta.url)))
 
-/**
- * if package was linked to global strip relative path from output
- * @param {String} $0 raw value of argv.$0
- * @returns {String} script name with relative path stripped if linked
- */
-function getScriptName ($0) {
-  if ($0.startsWith('.')) {
-    return $0.substring($0.lastIndexOf('/') + 1)
-  }
-  return $0
+if (!isSupported(process.versions.node, engines.node)) {
+  console.error(
+    `Unsupported Node.js version: ${process.version}\n` +
+    `xst requires Node.js ${engines.node}.\n` +
+    'Please install a supported version to run xst, e.g. from https://nodejs.org/')
+  process.exit(1)
 }
-
-function showCompletionHelp (scriptName) {
-  console.log(`
-${bb('Install command completions for ZSH and BASH')}
-  ${scriptName} completion`)
-}
-
-function showExamples (scriptName) {
-  console.log(`
-${bb('Examples:')}
-  ${scriptName} run 'count(//p)'
-  ${scriptName} list --tree --depth 1 /db/apps
-  ${scriptName} package install from-registry demo-apps
-`)
-}
-
-function showLogo () {
-  console.log('\n' +
-    ma`  ╲ ╱  ╓───  ──┰──` + '\n' +
-    rb`   ╳   ╰───╮   │  ` + '\n' +
-    yl`  ╱ ╲  ▂▁▁▁│   ┇  ` + '\n'
-  )
-  console.log(yl`A modern command line interface for exist-db`)
-}
-
-const parser = yargs(hideBin(process.argv))
-  .scriptName('xst')
-  .config('config', 'Read configuration file', configure)
-  .middleware(readConnection)
-  .usageConfiguration({ 'hide-types': true })
-  .completion('completion', false)
-  .strictCommands(true)
-  .strictOptions(false)
-  .help()
-  .command('$0 [<command>]', 'Interact with an eXist-db', () => {}, async (argv) => {
-    if (argv.command) {
-      console.error(`Command "${argv.command}" not recognized.
-Try xst --help`)
-    }
-
-    showLogo()
-    const scriptName = getScriptName(argv.$0)
-    showCompletionHelp(scriptName)
-    // append examples
-    showExamples(scriptName)
-  })
-  .command(commands)
-  // .recommendCommands()
-  .fail(false)
-
-parser.wrap(parser.terminalWidth())
 
 try {
-  await parser.parse()
+  await import('./app.js')
 } catch (error) {
-  handleError(error)
-  parser.getHelp()
-  process.exit(1)
+  // Belt and braces: a SyntaxError raised while loading the import chain
+  // points at an incompatible runtime as well (e.g. exotic setups where
+  // the reported version passes the check above).
+  if (error instanceof SyntaxError) {
+    console.error(
+      `${error.message}\n` +
+      `xst requires Node.js ${engines.node} (you are running ${process.version}).`)
+    process.exit(1)
+  }
+  throw error
 }
